Avoid importing metadata module in autocompletion
+ move get_runnable_pip to ``utils.misc`` module
to avoid importing ``build_env`` module.
